requirements.txt
arrow==1.2.3
certifi==2024.2.2
charset-normalizer==2.1.1
docker==6.0.0
idna==3.7
numpy==1.21.6
packaging==24.0
pandas==1.3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.28.1
six==1.16.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==1.26.18
websocket-client==1.6.1
